4-RAG+Document+Q&A/requirements.txt
langchain
python-dotenv
streamlit
langchain_community
langchain-groq
pypdf
faiss-cpu
langchain_chroma